Promise now returned when server is launched; finished example API docs.
