+++ package.json
@@ -9,44 +9,25 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/anyweez/simplespeak.git"
+    "url": "git+https://github.com/anyweez/harold.git"
   },
   "author": "Luke Segars (@anyweez)",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/anyweez/simplespeak/issues"
+    "url": "https://github.com/anyweez/harold/issues"
   },
-  "homepage": "https://github.com/anyweez/simplespeak#readme",
+  "homepage": "https://github.com/anyweez/harold#readme",
   "dependencies": {
     "colors": "^1.1.2",
-    "feed-read": "0.0.1",
-    "gulp": "^3.9.1",
-    "gulp-css-validator": "0.0.4",
-    "gulp-htmlhint": "^0.3.1",
-    "gulp-js-obfuscator": "^1.0.0",
-    "gulp-jslint": "^0.2.2",
-    "gulp-uglify": "^1.5.3",
     "hapi": "^13.2.1",
-    "hapi-passport": "0.0.6",
-    "hapi-plugin-websocket": "^0.9.1",
-    "hapi-swagger": "^6.2.0",
     "inert": "^3.2.1",
     "joi": "^8.0.4",
     "mz": "^2.4.0",
-    "passport": "^0.3.2",
-    "passport-github2": "^0.1.10",
     "path": "^0.12.7",
-    "request": "^2.69.0",
-    "vision": "^4.1.0",
-    "ws": "^1.1.0"
-  },
-  "devDependencies": {
-    "babel-preset-es2015": "^6.9.0",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^6.1.2"
+    "vision": "^4.1.0"
   },
   "apidoc": {
-    "title": "Custom apiDoc browser title",
-    "url": "https://api.github.com/v1"
+    "title": "Harold API docs",
+    "url": "https://yourdomain.com"
   }
 }
