+++ routes/get/chats.js
@@ -1,5 +1,14 @@
+/**
+ * @api {get} /chats Get list of all available chat messages
+ * @apiName GetChats
+ * @apiGroup Chats
+ * 
+ * @apiSuccess {Array} chats    All available chat messages
+ */
 module.exports = {
     handler(request, reply, state) {
-        reply(state.chatlog.getAll());
+        reply({
+            chats: state.chatlog.getAll(),
+        });
     },
 };
