+++ routes/post/chats.js
@@ -1,5 +1,16 @@
 const Boom = require('boom');
 
+/**
+ * @api {post} /chats Record new chat message
+ * @apiName PostChat
+ * @apiGroup Chats
+ * @apiDescription Adds a new message to the global chat log. Messages can be retrieved via the GET /chats
+ * endpoint. 
+ * 
+ * @apiParam {string}   from    User who sent the message
+ * @apiParam {string}   message Message content (raw text)
+ * @apiSuccess {Array}  chats   All available chat messages
+ */
 module.exports = {
     handler(request, reply, state) {
         if (!request.payload.from) reply(Boom.badRequest('Must specify a user name'));
@@ -10,7 +21,9 @@ module.exports = {
                 message: request.payload.message,
             });
 
-            reply(state.chatlog.getAll());
+            reply({
+                chats: state.chatlog.getAll()
+            });
         }
     },
 };
